Finish exercise 07 of the flexbox directory
I completed the layout until a certain point, then I used devtools to
debug the solution.html in order to fix up some things. I was not aware
you could center text within a div using flexbox. The width was also a
factor when it came to flex-wrap which I didn't understand.